% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publications.R
\name{get_publications}
\alias{get_publications}
\title{Get Publications from Google Scholar and ORCID}
\usage{
get_publications(orcid_id, scholar_id)
}
\arguments{
\item{orcid_id}{ORCID ID}

\item{scholar_id}{Google Scholar ID}
}
\value{
A tibble of combined and de-duplicated publications
}
\description{
Get Publications from Google Scholar and ORCID
}
