% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_pub.R
\name{get_publications_from_orcid}
\alias{get_publications_from_orcid}
\title{Get Publications from ORCID}
\usage{
get_publications_from_orcid(orcid_id)
}
\arguments{
\item{orcid_id}{ORCID ID}
}
\value{
A tibble of publications from ORCID
}
\description{
Get Publications from ORCID
}
