% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_pub.R
\name{get_publications_from_scholar}
\alias{get_publications_from_scholar}
\title{Get Publications from Google Scholar}
\usage{
get_publications_from_scholar(scholar_id)
}
\arguments{
\item{scholar_id}{Google Scholar ID}
}
\value{
A tibble of publications from Google Scholar
}
\description{
Get Publications from Google Scholar
}
